Change storage in PoolConfig to uint256 for contract size consideration

+++ contracts/BasePool.sol
@@ -66,7 +66,7 @@ abstract contract BasePool is BasePoolStorage, OwnableUpgradeable, ILiquidityPro
         _humaConfig = humaConfig;
         _feeManagerAddress = feeManager;
 
-        _poolConfig._withdrawalLockoutPeriodInSeconds = uint64(SECONDS_IN_180_DAYS);
+        _poolConfig._withdrawalLockoutPeriodInSeconds = SECONDS_IN_180_DAYS;
         _poolConfig._poolDefaultGracePeriodInSeconds = HumaConfig(humaConfig)
             .protocolDefaultGracePeriod();
         _status = PoolStatus.Off;
@@ -223,7 +223,7 @@ abstract contract BasePool is BasePoolStorage, OwnableUpgradeable, ILiquidityPro
     function setAPR(uint256 aprInBps) external virtual override {
         onlyOwnerOrHumaMasterAdmin();
         require(aprInBps <= 10000, "INVALID_APR");
-        _poolConfig._poolAprInBps = uint16(aprInBps);
+        _poolConfig._poolAprInBps = aprInBps;
         emit APRUpdated(aprInBps);
     }
 
@@ -234,7 +234,7 @@ abstract contract BasePool is BasePoolStorage, OwnableUpgradeable, ILiquidityPro
     function setReceivableRequiredInBps(uint256 receivableInBps) external virtual override {
         onlyOwnerOrHumaMasterAdmin();
         require(receivableInBps <= 10000, "INVALID_COLLATERAL_IN_BPS");
-        _poolConfig._receivableRequiredInBps = uint16(receivableInBps);
+        _poolConfig._receivableRequiredInBps = receivableInBps;
     }
 
     /**
@@ -244,7 +244,7 @@ abstract contract BasePool is BasePoolStorage, OwnableUpgradeable, ILiquidityPro
     function setMaxCreditLine(uint256 maxCreditLine) external virtual override {
         onlyOwnerOrHumaMasterAdmin();
         require(maxCreditLine > 0, "MAX_IS_ZERO");
-        _poolConfig._maxCreditLine = uint96(maxCreditLine);
+        _poolConfig._maxCreditLine = maxCreditLine;
     }
 
     /**
@@ -272,9 +272,7 @@ abstract contract BasePool is BasePoolStorage, OwnableUpgradeable, ILiquidityPro
      */
     function setPoolDefaultGracePeriod(uint256 gracePeriodInDays) external virtual override {
         onlyOwnerOrHumaMasterAdmin();
-        _poolConfig._poolDefaultGracePeriodInSeconds = uint64(
-            gracePeriodInDays * SECONDS_IN_A_DAY
-        );
+        _poolConfig._poolDefaultGracePeriodInSeconds = gracePeriodInDays * SECONDS_IN_A_DAY;
         emit PoolDefaultGracePeriodChanged(gracePeriodInDays, msg.sender);
     }
 
@@ -284,9 +282,7 @@ abstract contract BasePool is BasePoolStorage, OwnableUpgradeable, ILiquidityPro
      */
     function setWithdrawalLockoutPeriod(uint256 lockoutPeriodInDays) external virtual override {
         onlyOwnerOrHumaMasterAdmin();
-        _poolConfig._withdrawalLockoutPeriodInSeconds = uint64(
-            lockoutPeriodInDays * SECONDS_IN_A_DAY
-        );
+        _poolConfig._withdrawalLockoutPeriodInSeconds = lockoutPeriodInDays * SECONDS_IN_A_DAY;
         emit WithdrawalLockoutPeriodUpdated(lockoutPeriodInDays, msg.sender);
     }
 
@@ -296,7 +292,7 @@ abstract contract BasePool is BasePoolStorage, OwnableUpgradeable, ILiquidityPro
      */
     function setPoolLiquidityCap(uint256 liquidityCap) external virtual override {
         onlyOwnerOrHumaMasterAdmin();
-        _poolConfig._liquidityCap = uint96(liquidityCap);
+        _poolConfig._liquidityCap = liquidityCap;
         emit PoolLiquidityCapChanged(liquidityCap, msg.sender);
     }
